chat-with-plg-data/webhook_server/requirements.txt
fastapi==0.103.0
pydantic==1.10.7
python-dotenv==1.0.0
SQLAlchemy==2.0.15
uvicorn==0.23.2
mysqlclient==2.1.1